Update to float

+++ methods.py
@@ -2,15 +2,15 @@ from matplotlib import pyplot as plt
 import numpy as np
 
 def dist(qg,qn):
-	distancia = (int(qg[0])-int(qn[0]))**2 + (int(qg[1])-int(qn[1]))**2#Cálculo da distância entre dois pontos, sem tirar a raiz, pelo fato do cálculo da força ter a distância ao quadrado, assim cortando a raiz
+	distancia = (float(qg[0])-float(qn[0]))**2 + (float(qg[1])-float(qn[1]))**2#Cálculo da distância entre dois pontos, sem tirar a raiz, pelo fato do cálculo da força ter a distância ao quadrado, assim cortando a raiz
 	return distancia
 
 def field(qg):
 	k = 9*1e9
-	return k*int(qg)#Cálculo do campo em relação a carga geradora, sem a distância, pois ela varia para cada carga
+	return k*float(qg)#Cálculo do campo em relação a carga geradora, sem a distância, pois ela varia para cada carga
 
 def forca(campo,qn,distancia):
 	if distancia == 0:#Verifica se a carga qn e a geradora de campo estão na mesma posição
 		print('Erro: Duas cargas não podem ocupar a mesma posição')
 		exit()
-	return campo*int(qn)/distancia#Cálculo da força entre a carga geradora e cada carga qn
+	return campo*float(qn)/distancia#Cálculo da força entre a carga geradora e cada carga qn
